Issue #86
